enable /W4 warnings on visual studio.  clear up some of the new warnings.  move some more options to be HELICS_XXXX, start adding the docs of the options.  Make sure the code uses the new option names instead of the old ones.

+++ src/helics/application_api/Federate.cpp
@@ -418,7 +418,7 @@ void Federate::setProperty (int32_t option, Time timeValue)
 
 void Federate::setProperty (int32_t option, int32_t optionValue)
 {
-    coreObject->setIntegerProperty (fedID, option, optionValue);
+    coreObject->setIntegerProperty (fedID, option, static_cast<int16_t> (optionValue));
 }
 
 Time Federate::getTimeProperty (int32_t option) const { return coreObject->getTimeProperty (fedID, option); }
